Aula7 sobre HashMap

+++ src/br/com/zup/Aula6/Aula6_Exercicio2_ListaComplementar.java
@@ -6,17 +6,17 @@ public class Aula6_Exercicio2_ListaComplementar {
     public static void main(String[] args) {
         Scanner leitor = new Scanner(System.in);
         System.out.println("Digite quantidade de números que deseja calcular a média: ");
-        double quantidadeNumeros = leitor.nextDouble();
+        int quantidadeNumeros = leitor.nextInt();
         double media = 0;
         double soma = 0;
-        double numeroInserido = 0;
+        int numeroInserido = 0;
 
 
         int contador = 0;
 
         while (contador < quantidadeNumeros) {
             System.out.println("Digite o número " + contador + ":");
-            numeroInserido = leitor.nextDouble();
+            numeroInserido = leitor.nextInt();
             soma = soma + numeroInserido;
             contador++;
